+++ crates/forge-runner/src/lib.rs
@@ -1,4 +1,3 @@
-use crate::build_trace_data::test_sierra_program_path::VersionedProgramPath;
 use crate::coverage_api::run_coverage;
 use crate::forge_config::{ExecutionDataToSave, ForgeConfig, TestRunnerConfig};
 use crate::fuzzer::RandomFuzzer;
@@ -7,13 +6,11 @@ use crate::test_case_summary::TestCaseSummary;
 use anyhow::{anyhow, Result};
 use build_trace_data::save_trace_data;
 use cairo_lang_sierra::program::{ConcreteTypeLongId, Function, TypeDeclaration};
-use camino::Utf8Path;
+use camino::Utf8PathBuf;
 use cheatnet::runtime_extensions::forge_config_extension::config::RawFuzzerConfig;
 use futures::stream::FuturesUnordered;
 use futures::StreamExt;
-use package_tests::with_config_resolved::{
-    TestCaseWithResolvedConfig, TestTargetWithResolvedConfig,
-};
+use package_tests::with_config_resolved::TestCaseWithResolvedConfig;
 use profiler_api::run_profiler;
 use shared::print::print_as_warning;
 use std::collections::HashMap;
@@ -68,9 +65,10 @@ pub fn maybe_save_trace_and_profile(
     }) = result
     {
         if execution_data_to_save.is_vm_trace_needed() {
-            let trace_path = save_trace_data(name, trace_data)?;
+            let name = sanitize_filename::sanitize(name.replace("::", "_"));
+            let trace_path = save_trace_data(&name, trace_data)?;
             if execution_data_to_save.profile {
-                run_profiler(name, &trace_path, &execution_data_to_save.additional_args)?;
+                run_profiler(&name, &trace_path, &execution_data_to_save.additional_args)?;
             }
             return Ok(Some(trace_path));
         }
@@ -95,33 +93,13 @@ pub fn maybe_generate_coverage(
     Ok(())
 }
 
-pub fn maybe_save_versioned_program(
-    execution_data_to_save: &ExecutionDataToSave,
-    test_target: &TestTargetWithResolvedConfig,
-    versioned_programs_dir: &Utf8Path,
-    package_name: &str,
-) -> Result<Option<VersionedProgramPath>> {
-    let maybe_versioned_program_path = if execution_data_to_save.is_vm_trace_needed() {
-        Some(VersionedProgramPath::save_versioned_program(
-            &test_target.sierra_program.clone().into(),
-            test_target.tests_location,
-            versioned_programs_dir,
-            package_name,
-        )?)
-    } else {
-        None
-    };
-
-    Ok(maybe_versioned_program_path)
-}
-
 #[must_use]
 pub fn run_for_test_case(
     args: Vec<ConcreteTypeLongId>,
     case: Arc<TestCaseWithResolvedConfig>,
     casm_program: Arc<AssembledProgramWithDebugInfo>,
     forge_config: Arc<ForgeConfig>,
-    maybe_versioned_program_path: Arc<Option<VersionedProgramPath>>,
+    versioned_program_path: Arc<Utf8PathBuf>,
     send: Sender<()>,
 ) -> JoinHandle<Result<AnyTestCaseSummary>> {
     if args.is_empty() {
@@ -130,7 +108,7 @@ pub fn run_for_test_case(
                 case,
                 casm_program,
                 forge_config.test_runner_config.clone(),
-                maybe_versioned_program_path,
+                versioned_program_path,
                 send,
             )
             .await?;
@@ -143,7 +121,7 @@ pub fn run_for_test_case(
                 case,
                 casm_program,
                 forge_config.test_runner_config.clone(),
-                maybe_versioned_program_path,
+                versioned_program_path,
                 send,
             )
             .await??;
@@ -174,7 +152,7 @@ fn run_with_fuzzing(
     case: Arc<TestCaseWithResolvedConfig>,
     casm_program: Arc<AssembledProgramWithDebugInfo>,
     test_runner_config: Arc<TestRunnerConfig>,
-    maybe_versioned_program_path: Arc<Option<VersionedProgramPath>>,
+    versioned_program_path: Arc<Utf8PathBuf>,
     send: Sender<()>,
 ) -> JoinHandle<Result<TestCaseSummary<Fuzzing>>> {
     tokio::task::spawn(async move {
@@ -207,7 +185,7 @@ fn run_with_fuzzing(
                 case.clone(),
                 casm_program.clone(),
                 test_runner_config.clone(),
-                maybe_versioned_program_path.clone(),
+                versioned_program_path.clone(),
                 send.clone(),
                 fuzzing_send.clone(),
             ));