+++ crates/forge-runner/src/package_tests/raw.rs
@@ -5,7 +5,7 @@ use camino::Utf8PathBuf;
 /// these structs are representation of scarb output for `scarb build --test`
 /// produced by scarb
 pub struct TestTargetRaw {
-    pub sierra_file_path: Utf8PathBuf,
     pub sierra_program: ProgramArtifact,
+    pub sierra_program_path: Utf8PathBuf,
     pub tests_location: TestTargetLocation,
 }