+++ crates/forge-runner/src/running/with_config.rs
@@ -40,7 +40,7 @@ pub fn test_target_with_config(test_target_raw: TestTargetRaw) -> Result<TestTar
     let type_declarations = by_id!(type_declarations);
 
     let casm_program = Arc::new(compile_sierra_at_path(
-        &test_target_raw.sierra_file_path,
+        &test_target_raw.sierra_program_path,
         &SierraType::Raw,
     )?);
 
@@ -81,6 +81,7 @@ pub fn test_target_with_config(test_target_raw: TestTargetRaw) -> Result<TestTar
         tests_location: test_target_raw.tests_location,
         test_cases,
         sierra_program: test_target_raw.sierra_program,
+        sierra_program_path: test_target_raw.sierra_program_path.into(),
         casm_program,
     })
 }