+++ crates/forge/tests/integration/store_load.rs
@@ -556,12 +556,13 @@ fn store_load_felt_to_felt() {
     assert_passed(&result);
 }
 
+#[cfg(not(target_os = "windows"))]
 #[test]
 fn fork_store_load() {
     let test = test_utils::test_case!(formatdoc!(
         r#"
             use starknet::{{ ContractAddress, contract_address_const }};
-            use snforge_std::{{ BlockTag, BlockId, load, store }};
+            use snforge_std::{{ load, store }};
 
             #[starknet::interface]
             trait IHelloStarknet<TContractState> {{