+++ crates/sncast/tests/helpers/devnet.rs
@@ -34,7 +34,13 @@ fn start_devnet() {
         }
     }
 
-    Command::new("tests/utils/devnet/starknet-devnet")
+    let devnet_path = if cfg!(target_os = "windows") {
+        "tests/utils/devnet/bin/starknet-devnet.exe"
+    } else {
+        "tests/utils/devnet/starknet-devnet"
+    };
+
+    Command::new(devnet_path)
         .args([
             "--port",
             &port,
@@ -81,11 +87,17 @@ fn start_devnet() {
 #[dtor]
 fn stop_devnet() {
     let port = Url::parse(URL).unwrap().port().unwrap_or(80).to_string();
-    Command::new("pkill")
-        .args([
-            "-f",
-            &format!("starknet-devnet.*{}.*{}", &port, &SEED.to_string())[..],
-        ])
-        .spawn()
-        .expect("Failed to kill devnet processes");
+    let pattern = format!("starknet-devnet.*{port}.*{SEED}");
+
+    if cfg!(target_os = "windows") {
+        Command::new("taskkill")
+            .args(["/IM", &pattern, "/F"])
+            .output()
+            .expect("Failed to kill devnet processes");
+    } else {
+        Command::new("pkill")
+            .args(["-f", &pattern])
+            .output()
+            .expect("Failed to kill devnet processes");
+    }
 }